tools/picturas-rotate-tool-ms/pyproject.toml
[tool.poetry]
name = "picturas-rotate-tool-ms"
version = "0.1.0"
description = "PictuRAS microservice (tool) that rotates images."
authors = ["Marta Rodrigues <pg55982@alunos.uminho.pt>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
pika = "^1.3.2"
pydantic = "^2.10.3"
pillow = "^11.0.0"

[tool.poetry.group.dev]
dependencies = { yapf = "^0.43.0" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
